remove BUILD_ALWAYS for external dependencies
Recently the default for Build_Deps was changed from IfNotFound to Always which
makes BUILD_ALWAYS redundant.  In particular, there is no way for users to
override an explicit BUILD_ALWAYS on a dependency which makes it impossible to
build TRIQS packages in environments with restricted network access or without
Git.